backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0

pandas==2.2.2
scikit-learn==1.4.2
numpy==1.26.4
onnxruntime==1.17.1

pillow==10.3.0
requests==2.31.0
itsdangerous==2.2.0
